added sanitize middleware
